Gjøre alle kort dragbare og bruke div som lenke for styling i andre nettlesere
