+++ src/moduler/aktivitet/aktivitet-kort/aktivitetskort.less
@@ -84,8 +84,6 @@
 
 .aktivitetskort--flyttbar {
     cursor: move;
-    cursor: -webkit-grab; // lesshint duplicateProperty: false
-    cursor: -moz-grab; // lesshint duplicateProperty: false
     cursor: grab; // lesshint duplicateProperty: false
 }
